Move SentenceTransformerWithTruncate to a separate module to simplify imports. Imports are still only done when needed.
Don't override `encode()`, rather rename to `encode_with_truncate()`.

Now under normal usage a the model will be a SentenceTransformerWithTruncate, but if the enduser
does weird loading and the model is a standard SentenceTransformer the standard `encode()` will be called.

Some typing improvements.
